+++ src/utils/bn254_circom_x5_5.rs
@@ -1,8 +1,8 @@
 use crate::poseidon::sbox::PoseidonSbox;
 
-pub const FULL_ROUNDS: usize = 8;
-pub const PARTIAL_ROUNDS: usize = 60;
-pub const WIDTH: usize = 5;
+pub const FULL_ROUNDS: u8 = 8;
+pub const PARTIAL_ROUNDS: u8 = 60;
+pub const WIDTH: u8 = 5;
 pub const SBOX: PoseidonSbox = PoseidonSbox::Exponentiation(5);
 pub const ROUND_CONSTS: [&str; 340] = [
 	"0x0eb544fee2815dda7f53e29ccac98ed7d889bb4ebd47c3864f3c2bd81a6da891",
